+++ grails-app/controllers/com/cc/content/blog/BlogController.groovy
@@ -54,10 +54,10 @@ class BlogController {
         int defaultMax = grailsApplication.config.cc.plugins.content.blog.list.max ?: 10
 
         params.offset = offset ? offset: 0
-        params.max = Math.min(max ?: 2, 100)
+        params.max = Math.min(max ?: defaultMax, 100)
 
         StringBuilder query = new StringBuilder("""SELECT new Map(b.id as id, b.body as body, b.title as title,
-                            b.subTitle as subTitle, b.author as author, b.dateCreated as dateCreated) FROM Blog b """)
+                            b.subTitle as subTitle, b.author as author, b.publishedDate as publishedDate) FROM Blog b """)
 
         if(tag) {
             query.append(", ${TagLink.class.name} tagLink WHERE b.id = tagLink.tagRef ")